requirements.txt
Pyrogram
TgCrypto
uvloop
lexica-api
httpx[http2]
python-dotenv
pony
pillow